rename the files
